BlogController 410 errors

+++ backend/src/blog/blog.controller.ts
@@ -13,6 +13,7 @@ import {
 import { BlogService } from "./blog.service";
 import { BlogSchema, CreateBlogSchema, UpdateBlogSchema } from "@repo/common";
 import { ZodValidationPipe } from "../common/pipes/zod.validation.pipe";
+import { ResourceGoneException } from "src/common/exceptions";
 import { BlogDto, CreateBlogDto, UpdateBlogDto } from "../dto/dto";
 import {
   ApiBody,
@@ -52,7 +53,11 @@ export class BlogController {
   async getBlogById(
     @Param("blogId", ParseIntPipe) blogId: number,
   ): Promise<BlogDto> {
-    return await this.blogService.getBlogById(blogId);
+    const blog = await this.blogService.getBlogById(blogId);
+    if (!blog) {
+      throw new ResourceGoneException(`Blog ${blogId} not found.`);
+    }
+    return blog;
   }
 
   /**
@@ -88,7 +93,11 @@ export class BlogController {
     @Param("blogId", ParseIntPipe) blogId: number,
     @Body(new ZodValidationPipe(BlogSchema)) blog: BlogDto,
   ): Promise<BlogDto> {
-    return await this.blogService.replaceBlog(blogId, blog);
+    const updated = await this.blogService.replaceBlog(blogId, blog);
+    if (!updated) {
+      throw new ResourceGoneException(`Cannot replace: Blog ${blogId} does not exist`);
+    }
+    return updated;
   }
 
   /**
@@ -107,7 +116,11 @@ export class BlogController {
     @Param("blogId", ParseIntPipe) blogId: number,
     @Body(new ZodValidationPipe(UpdateBlogSchema)) blog: UpdateBlogDto,
   ): Promise<BlogDto> {
-    return await this.blogService.modifyBlog(blogId, blog);
+    const updated = await this.blogService.modifyBlog(blogId, blog);
+    if (!updated) {
+      throw new ResourceGoneException(`Cannot modify: Blog ${blogId} does not exist`);
+    }
+    return updated;
   }
 
   /**
@@ -123,6 +136,10 @@ export class BlogController {
   async deleteBlog(
     @Param("blogId", ParseIntPipe) blogId: number,
   ): Promise<void> {
+    const blog = await this.blogService.getBlogById(blogId);
+    if (!blog) {
+      throw new ResourceGoneException(`Cannot delete: Blog ${blogId} does not exist`);
+    }
     return await this.blogService.deleteBlog(blogId);
   }
 }
